Add DEMO_RUN_SPEED
This allows us to change the speed of the command during the demo
script.

When doing demos with user namespaces, it is common to show how an
exploit works without userns and then how it can't be exploited with
userns. For that we do the same steps with and without userns. To make
the demo feel not so slow and boring, it is nice to tune the speed a
little bit.

Also, now that we have desc_type(), sometimes the comments can be
displayed faster than the things we run (especially if they are long
comments, people can read fast when it is just text and not code).

But it's not always the case for desc_type() to be slower. Let's just
have a global var that long demos can use to tune the speed of different
parts.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>